refactor: move UMI methods to platform/illumina.py
